requirements.txt
# For Gemini API
google-generativeai

# For loading environment variables (like API keys from .env)
python-dotenv

# For handling YAML configuration files (game.cfg, logging_conf.yaml, etc.)
PyYAML

# SciPy was an import in the original student_agent.py stub.
# Include if it's used by the broader framework or if you plan to use it.
scipy

langgraph
langchain-google-genai

python-dotenv
typing_inspect
typing_extensions